rename components to jsx files
