+++ src/js/components/AppStatusContainer.jsx
@@ -47,8 +47,18 @@ export default class AppStatusContainer extends React.Component {
 
     return (
       <div id="app-status-container" className={showAppStatus}>
-        <img id="loading-articles" className={`img-responsive center-block ${showLoadingImg}`} src="https://res.cloudinary.com/worldgeek/image/upload/v1493551216/loading-img_zqnq20.gif" />
-        <img id="error-loading-articles" className={`img-responsive center-block ${showErrorImg}`} src="https://res.cloudinary.com/worldgeek/image/upload/v1493551216/error-loading-img_z8tzia.png" />
+        <img
+          id="loading-articles"
+          alt="Loading articles... Please wait..."
+          className={`img-responsive center-block ${showLoadingImg}`}
+          src="https://res.cloudinary.com/worldgeek/image/upload/v1493551216/loading-img_zqnq20.gif"
+        />
+        <img
+          id="error-loading-articles"
+          alt="Failed to load articles. Please try again."
+          className={`img-responsive center-block ${showErrorImg}`}
+          src="https://res.cloudinary.com/worldgeek/image/upload/v1493551216/error-loading-img_z8tzia.png"
+        />
         <h3 className="text-center">{statusMsg}</h3>
       </div>
     );
