+++ src/js/components/WelcomeMessageContainer.jsx
